cepheus: Install vendorcompat verison of protobuf
Some prebuilt vendor modules contain references to
libprotobuf-cpp-*.so, but the interface is not stable.  Upgrading
protobuf would cause those modules to fail, so the vendor version
of the new protobuf library is renamed to libprotobuf-cpp-*-3.9.1.so.
Manually install old libprotobuf-cpp-*.so files to /vendor
to avoid breaking products that have prebuilts that reference them.

Once the new version of protobuf is in each product can be
inspected for references to libprotobuf-cpp-*.so on /vendor and
these packages can be included only on those devices that need
them.

Bug: 117607748
Test: m checkbuild
Change-Id: I8ac955eb703e3faf22ff930c59b30385f374ad0a

The entry in public.libraries.txt is needed to allow the
linker to find and use the lib outside the VNDK.

Signed-off-by: Carlos Ayrton Lopez Arroyo <15030201@itcelaya.edu.mx>